+++ src/popups/open-translation.tsx
@@ -8,7 +8,7 @@ import { DialogFooter } from "@/components/ui/dialog";
 import LangSelector from "@/components/lang-selector";
 import FilePicker from "@/components/fields/file-picker";
 import FileActions from "@/actions/file";
-import { PopupFormProps } from "@/lib/types";
+import { ILangInputState, PopupFormProps } from "@/lib/types";
 import { useState, useTransition } from "react";
 import LoadingButton from "@/components/loading-button";
 import { FolderOpen } from "lucide-react";
@@ -17,11 +17,14 @@ import { toast } from "sonner";
 import { useAppTranslation } from "@/context/translation";
 import { detectLanguageCode, getFileName, getFormatFromPath } from "@/lib/helpers";
 import RecentTranslations from "@/lib/store/recent-translations";
+import { useGlossary } from "@/context/glossary-sidebar";
+import GlossaryActions from "@/lib/store/glossary";
 
 export default function OpenTranslationPopup({triggerButton}: PopupFormProps){
      const [isOpening, startTransition] = useTransition()
      const [open, setOpen] = useState(false)
      const {setTable, updateLangs, setFiles, setBaseKeys, setIsDirty} = useAppTranslation()
+     const {setGlossary} = useGlossary()
      const form = useForm<OpenTranslationType>({
           resolver: zodResolver(OpenTranslationSchema),
           defaultValues: {
@@ -51,10 +54,13 @@ export default function OpenTranslationPopup({triggerButton}: PopupFormProps){
                               targetPath: values.targetPath,
                               format: await getFormatFromPath(values.basePath)
                          })
-                         updateLangs({
+                         const langs: ILangInputState = {
                               base: values.baseLang,
-                              target: values.targetLang,
-                         }),
+                              target: values.targetLang
+                         }
+                         updateLangs(langs)
+                         const glossary = await GlossaryActions.getGlossary(langs)
+                         setGlossary(glossary)
                          setFiles({
                               basePath: values.basePath,
                               targetPath: values.targetPath,