+++ src/schemas/types.ts
@@ -1,5 +1,5 @@
 import * as z from "zod"
-import { BatchRenameKeysSchema, FindSchema, NewTranslationSchema, OpenTranslationSchema, OpenXliffSchema, ReplaceTranslationSchema, SpellCheckSchema, TransliterateScriptSchema } from "."
+import { AddToGlossarySchema, BatchRenameKeysSchema, FindSchema, NewTranslationSchema, OpenTranslationSchema, OpenXliffSchema, ReplaceTranslationSchema, SpellCheckSchema, TransliterateScriptSchema } from "."
 import { AutoTranslateSchema, providerField } from "./auto-translate"
 
 export type NewTranslationType = z.infer<typeof NewTranslationSchema>
@@ -12,3 +12,4 @@ export type AutoTranslateProvider = z.infer<typeof providerField>
 export type TransliterateScriptType = z.infer<typeof TransliterateScriptSchema>
 export type SpellCheckType = z.infer<typeof SpellCheckSchema>
 export type OpenXliffType = z.infer<typeof OpenXliffSchema>
+export type AddToGlossaryType = z.infer<typeof AddToGlossarySchema>